fix: refactor article code
