+++ app/src/main/java/com/example/fitfoood/adapter/ArticleListAdapter.kt
@@ -7,21 +7,37 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.example.fitfoood.R
 import com.example.fitfoood.data.response.ArticleItem
-import com.example.fitfoood.databinding.Artikel2RowBinding
-import com.example.fitfoood.view.artikel.DetailArtikelActivity
+import com.example.fitfoood.databinding.ArticleRowListBinding
+import com.example.fitfoood.utils.Constants
+import com.example.fitfoood.view.article.ArticleDetailActivity
 
+/**
+ * Article List Adapter
+ *
+ * This adapter is used to show article data in list view
+ *
+ * @property listItem list of article data
+ * @see ArticleItem
+ */
 class ArticleListAdapter(private var listItem: List<ArticleItem>) : RecyclerView.Adapter<ArticleListAdapter.ViewHolder>() {
 
-    class ViewHolder(private val binding: Artikel2RowBinding) : RecyclerView.ViewHolder(binding.root) {
-        fun bind(artikel: ArticleItem) {
-            binding.tvName.text = artikel.title
+    /**
+     * ViewHolder class
+     *
+     * This class is used to bind article data to the view
+     *
+     * @param binding article row list binding
+     */
+    class ViewHolder(private val binding: ArticleRowListBinding) : RecyclerView.ViewHolder(binding.root) {
+        fun bind(article: ArticleItem) {
+            binding.tvName.text = article.title
 
             Glide.with(binding.root)
-                .load(artikel.thumbnail)
+                .load(article.thumbnail)
                 .into(binding.imgPhoto)
 
-            val categoryIcon = when (artikel.category) {
-                "hidup sehat" -> R.drawable.icon_hidupsehat
+            val categoryIcon = when (article.category) {
+                "hidup-sehat" -> R.drawable.icon_hidupsehat
                 "olahraga" -> R.drawable.icon_olahraga
                 else -> R.drawable.icon_hidupsehat // Optional default icon
             }
@@ -29,22 +45,40 @@ class ArticleListAdapter(private var listItem: List<ArticleItem>) : RecyclerView
         }
     }
 
+    /**
+     * Create ViewHolder
+     *
+     * @param parent parent view group
+     * @param viewType view type
+     * @return ViewHolder
+     */
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val binding = Artikel2RowBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        val binding = ArticleRowListBinding.inflate(LayoutInflater.from(parent.context), parent, false)
         return ViewHolder(binding)
     }
 
+    /**
+     * Get item count
+     *
+     * @return item count
+     */
     override fun getItemCount(): Int {
         return listItem.size
     }
 
+    /**
+     * Bind data to the view
+     *
+     * @param holder view holder
+     * @param position item position
+     */
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val artikel = listItem[position]
         holder.bind(artikel)
 
         holder.itemView.setOnClickListener {
-            val intent = Intent(holder.itemView.context, DetailArtikelActivity::class.java)
-            intent.putExtra("Artikel", artikel)
+            val intent = Intent(holder.itemView.context, ArticleDetailActivity::class.java)
+            intent.putExtra(Constants.EXTRA_ARTICLE, artikel)
             holder.itemView.context.startActivity(intent)
         }
     }
