+++ app/src/main/java/com/example/fitfoood/view/article/ArticleActivity.kt
@@ -1,5 +1,6 @@
-package com.example.fitfoood.view.artikel
+package com.example.fitfoood.view.article
 
+import android.annotation.SuppressLint
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
@@ -9,25 +10,25 @@ import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.example.fitfoood.adapter.ArticleCardAdapter
 import com.example.fitfoood.data.ApiResponse
 import com.example.fitfoood.data.response.ArticleItem
-import com.example.fitfoood.databinding.ActivityArtikelBinding
+import com.example.fitfoood.databinding.ActivityArticleBinding
 import com.example.fitfoood.view.ViewModelFactory
 import com.example.fitfoood.view.main.HomeViewModel
 import com.google.android.material.tabs.TabLayoutMediator
 
 
-class ArtikelActivity : AppCompatActivity() {
+class ArticleActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityArticleBinding
+    private lateinit var viewModel: HomeViewModel
+    private var tabLayoutMediator: TabLayoutMediator? = null
 
     private val tabTitles = arrayOf("All", "Hidup Sehat", "Olahraga")
-
-    private lateinit var binding: ActivityArtikelBinding
-    private lateinit var viewModel: HomeViewModel
     private var articles: List<ArticleItem> = listOf()
-    private var tabLayoutMediator: TabLayoutMediator? = null
     private lateinit var label: String
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityArtikelBinding.inflate(layoutInflater)
+        binding = ActivityArticleBinding.inflate(layoutInflater)
         setContentView(binding.root)
         viewModel = ViewModelFactory.getInstance(this).create(HomeViewModel::class.java)
 
@@ -39,7 +40,13 @@ class ArtikelActivity : AppCompatActivity() {
         }
     }
 
-
+    /**
+     * Fetch articles from the API
+     *
+     * This function will fetch articles from the API and filter them based on the label.
+     *
+     * @see HomeViewModel.getArticles
+     */
     private fun fetchArticles() {
         viewModel.getArticles().observe(this) { articles ->
             when (articles) {
@@ -50,16 +57,20 @@ class ArtikelActivity : AppCompatActivity() {
                     showRecyclerList()
                 }
                 is ApiResponse.Error -> {
-                    // Handle error
+                    // TODO: Handle error
                 }
                 is ApiResponse.Loading -> {
-                    // Show loading
+                    // TODO: Show loading
                 }
             }
         }
     }
 
-
+    /**
+     * Setup the ViewPager
+     *
+     * This function will setup the ViewPager and the TabLayout for the articles.
+     */
     private fun setupViewPager() {
         if (tabLayoutMediator == null) {
             tabLayoutMediator = TabLayoutMediator(
@@ -79,6 +90,11 @@ class ArtikelActivity : AppCompatActivity() {
         }
     }
 
+    /**
+     * Show the recycler list
+     *
+     * This function will show the recycler list of articles.
+     */
     private fun showRecyclerList() {
         val adapter = ArticleCardAdapter(this.articles)
         with(binding.recyclerView) {
@@ -88,7 +104,20 @@ class ArtikelActivity : AppCompatActivity() {
         }
     }
 
+    /**
+     * SectionsPagerAdapter
+     *
+     * This class is used to manage the fragments in the ViewPager.
+     *
+     * @property articles list of articles
+     */
     inner class SectionsPagerAdapter(fm: FragmentManager, private var articles: List<ArticleItem>) : FragmentStateAdapter(fm, lifecycle) {
+
+        /**
+         * Create a new fragment
+         *
+         * This function will create a new fragment based on the position.
+         */
         override fun createFragment(position: Int): Fragment {
             val filteredArticles = when (position) {
                 0 -> articles // Show all articles for the "All" tab
@@ -99,10 +128,21 @@ class ArtikelActivity : AppCompatActivity() {
             return ContentFragment.newInstance(filteredArticles)
         }
 
+        /**
+         * Get the item count
+         *
+         * This function will return the number of tabs.
+         */
         override fun getItemCount(): Int {
             return tabTitles.size
         }
 
+        /**
+         * Update the articles
+         *
+         * This function will update the articles and notify the adapter.
+         */
+        @SuppressLint("NotifyDataSetChanged")
         fun updateArticles(newArticles: List<ArticleItem>) {
             articles = newArticles
             notifyDataSetChanged()
