+++ app/src/main/res/layout/activity_article_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.artikel.DetailArtikelActivity">
+    tools:context=".view.article.ArticleDetailActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
