+++ app/src/main/res/layout/article_row_card.xml
