+++ app/src/main/res/layout/custom_tab.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.fitfoood.view.artikel.ArtikelActivity"
+    tools:context="com.example.fitfoood.view.article.ArticleActivity"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     android:layout_margin="8dp"
