+++ app/src/main/res/layout/fragment_content.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.artikel.ContentFragment">
+    tools:context=".view.article.ContentFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerView2"
@@ -15,6 +15,6 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        tools:listitem="@layout/artikel2_row" />
+        tools:listitem="@layout/article_row_list" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
